requirements.txt
requests
ujson
twine
pycrypto